(chore) move NewList-related files to components/NewList subfolders
